Fix setup and teardown example
