+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `arduino_library_location.rb` script to print Arduino library location to stdout
 - `arduino_ci_remote.rb` now supports `--skip-unittests` and `--skip-compilation`.  If you skip both, only the `autolocate!` of the Arduino binary will be performed.
 - `keepachangelog_manager` gem to begin streamlining the release process
+- `unittest_setup()` and `unittest_teardown()` macros, my thanks to @hlovdal for contributing this code
 - Added rspec sensitivity to the environment variable `$ARDUINO_CI_SKIP_SPLASH_SCREEN_RSPEC_TESTS` (for `arduino_ci` gem hackers)
 
 ### Changed
